requirements.txt
langchain==0.2.16
langchain-community==0.2.17
langchain-openai==0.1.25
langgraph==0.2.28
chromadb==0.5.15
sentence-transformers==3.1.1
openai==1.51.0
rank-bm25==0.2.2
pypdf==4.3.1
beautifulsoup4==4.12.3
requests==2.32.3
python-docx==1.1.2
markdown==3.7
ragas==0.1.21
datasets==3.0.1
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.66.5
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.3
streamlit==1.38.0
